// apps/stroltm/ui/src/api/generated/api.ts
export interface ApiuResultSuccess {
  /**
   *
   * @type {string}
   * @memberof ApiuResultSuccess
   */
  data?: string;
}